+++ ndll/pipeline/workspace/device_workspace.cc
@@ -7,7 +7,7 @@ namespace ndll {
 
 template <>
 const TensorList<CPUBackend>& DeviceWorkspace::Input(int idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Input TensorList with given "
       "index does not have the calling backend type (CPUBackend)");
@@ -16,7 +16,7 @@ const TensorList<CPUBackend>& DeviceWorkspace::Input(int idx) const {
 
 template <>
 const TensorList<GPUBackend>& DeviceWorkspace::Input(int idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
@@ -25,7 +25,7 @@ const TensorList<GPUBackend>& DeviceWorkspace::Input(int idx) const {
 
 template <>
 TensorList<CPUBackend>* DeviceWorkspace::Output(int idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (CPUBackend)");
@@ -34,7 +34,7 @@ TensorList<CPUBackend>* DeviceWorkspace::Output(int idx) {
 
 template <>
 TensorList<GPUBackend>* DeviceWorkspace::Output(int idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
