+++ ndll/pipeline/workspace/host_workspace.cc
@@ -28,7 +28,7 @@ void HostWorkspace::GetSample(SampleWorkspace *ws,
 }
 
 int HostWorkspace::NumInputAtIdx(int idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   if (tensor_meta.first) {
     return cpu_inputs_[tensor_meta.second].size();
@@ -37,7 +37,7 @@ int HostWorkspace::NumInputAtIdx(int idx) const {
 }
 
 int HostWorkspace::NumOutputAtIdx(int idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   if (tensor_meta.first) {
     return cpu_outputs_[tensor_meta.second].size();
@@ -47,12 +47,12 @@ int HostWorkspace::NumOutputAtIdx(int idx) const {
 
 template <>
 const Tensor<CPUBackend>& HostWorkspace::Input(int idx, int data_idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Input with given index does not "
       "have the calling backend type (CPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       cpu_inputs_[tensor_meta.second].size());
 
   return *cpu_inputs_[tensor_meta.second][data_idx];
@@ -60,12 +60,12 @@ const Tensor<CPUBackend>& HostWorkspace::Input(int idx, int data_idx) const {
 
 template <>
 const Tensor<GPUBackend>& HostWorkspace::Input(int idx, int data_idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Input TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       gpu_inputs_[tensor_meta.second].size());
 
   return *gpu_inputs_[tensor_meta.second][data_idx];
@@ -73,12 +73,12 @@ const Tensor<GPUBackend>& HostWorkspace::Input(int idx, int data_idx) const {
 
 template <>
 Tensor<CPUBackend>* HostWorkspace::Output(int idx, int data_idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Output with given index does not "
       "have the calling backend type (CPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       cpu_outputs_[tensor_meta.second].size());
 
   return cpu_outputs_[tensor_meta.second][data_idx].get();
@@ -86,12 +86,12 @@ Tensor<CPUBackend>* HostWorkspace::Output(int idx, int data_idx) {
 
 template <>
 Tensor<GPUBackend>* HostWorkspace::Output(int idx, int data_idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       gpu_outputs_[tensor_meta.second].size());
 
   return gpu_outputs_[tensor_meta.second][data_idx].get();
