+++ ndll/pipeline/workspace/mixed_workspace.cc
@@ -6,7 +6,7 @@
 namespace ndll {
 
 int MixedWorkspace::NumInputAtIdx(int idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   if (tensor_meta.first) {
     return cpu_inputs_[tensor_meta.second].size();
@@ -16,12 +16,12 @@ int MixedWorkspace::NumInputAtIdx(int idx) const {
 
 template <>
 const Tensor<CPUBackend>& MixedWorkspace::Input(int idx, int data_idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Input with given index does not "
       "have the calling backend type (CPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       cpu_inputs_[tensor_meta.second].size());
 
   return *cpu_inputs_[tensor_meta.second][data_idx];
@@ -29,12 +29,12 @@ const Tensor<CPUBackend>& MixedWorkspace::Input(int idx, int data_idx) const {
 
 template <>
 const Tensor<GPUBackend>& MixedWorkspace::Input(int idx, int data_idx) const {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, input_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, input_index_map_.size());
   auto tensor_meta = input_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
 
-  NDLL_ENFORCE_VALID_INDEX((size_t)data_idx,
+  NDLL_ENFORCE_VALID_INDEX(data_idx,
       gpu_inputs_[tensor_meta.second].size());
 
   return *gpu_inputs_[tensor_meta.second][data_idx];
@@ -42,7 +42,7 @@ const Tensor<GPUBackend>& MixedWorkspace::Input(int idx, int data_idx) const {
 
 template <>
 TensorList<CPUBackend>* MixedWorkspace::Output(int idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (CPUBackend)");
@@ -51,7 +51,7 @@ TensorList<CPUBackend>* MixedWorkspace::Output(int idx) {
 
 template <>
 TensorList<GPUBackend>* MixedWorkspace::Output(int idx) {
-  NDLL_ENFORCE_VALID_INDEX((size_t)idx, output_index_map_.size());
+  NDLL_ENFORCE_VALID_INDEX(idx, output_index_map_.size());
   auto tensor_meta = output_index_map_[idx];
   NDLL_ENFORCE(!tensor_meta.first, "Output TensorList with given "
       "index does not have the calling backend type (GPUBackend)");
